+++ tests/Fixtures/Ordering/ADynamicController.php
@@ -8,7 +8,7 @@ use Karewan\KnRoute\Attributes\Get;
 
 class ADynamicController
 {
-	#[Get('/ordering/{value:any}')]
+	#[Get('/ordering/{value:segment}')]
 	public function dynamic(string $value): void
 	{
 		echo "dynamic:{$value}";